+++ bin/dialogs/preferencespanels/moodspanel.py
@@ -118,6 +118,9 @@ class MoodsPanel(wx.Panel):
         self.EnableArtistBackgrounds = wx.CheckBox(self, label='Enable artist backgrounds')
         self.EnableArtistBackgrounds.SetValue(str(artistBackgrounds.get('Enabled', 'False')).lower() == 'true')
         self.EnableArtistBackgrounds.Bind(wx.EVT_CHECKBOX, self.OnArtistBackgroundSettingsChanged)
+        self.UseCoverArtBackgrounds = wx.CheckBox(self, label='Use cover art / album art')
+        self.UseCoverArtBackgrounds.SetValue(str(artistBackgrounds.get('UseCoverArt', 'False')).lower() == 'true')
+        self.UseCoverArtBackgrounds.Bind(wx.EVT_CHECKBOX, self.OnArtistBackgroundSettingsChanged)
 
         self.MatchFieldDropdown = normalizeMacControlHeight(wx.ComboBox(self, value=artistBackgrounds.get('MatchField', '%AlbumArtist'), choices=self.matchFields, style=wx.CB_READONLY))
         self.MatchFieldDropdown.Bind(wx.EVT_COMBOBOX, self.OnArtistBackgroundSettingsChanged)
@@ -132,6 +135,8 @@ class MoodsPanel(wx.Panel):
         artistSettingsGrid = wx.FlexGridSizer(0, 2, 5, 10)
         artistSettingsGrid.AddGrowableCol(1, 1)
         artistSettingsGrid.Add(self.EnableArtistBackgrounds, 0, wx.ALIGN_CENTER_VERTICAL)
+        artistSettingsGrid.Add(wx.StaticText(self, -1, ''))
+        artistSettingsGrid.Add(self.UseCoverArtBackgrounds, 0, wx.ALIGN_CENTER_VERTICAL)
         artistSettingsGrid.Add(wx.StaticText(self, -1, ''))
         artistSettingsGrid.Add(wx.StaticText(self, -1, 'Default match field'), 0, wx.ALIGN_CENTER_VERTICAL)
         artistSettingsGrid.Add(self.MatchFieldDropdown, 0, wx.EXPAND)
@@ -201,6 +206,7 @@ class MoodsPanel(wx.Panel):
         self.BuildMoodList()
         artistBackgrounds = self.BeamSettings.getArtistBackgrounds()
         self.EnableArtistBackgrounds.SetValue(str(artistBackgrounds.get('Enabled', 'False')).lower() == 'true')
+        self.UseCoverArtBackgrounds.SetValue(str(artistBackgrounds.get('UseCoverArt', 'False')).lower() == 'true')
         self.MatchFieldDropdown.SetValue(artistBackgrounds.get('MatchField', '%AlbumArtist'))
         self.FallbackFieldDropdown.SetValue(artistBackgrounds.get('FallbackField', '%Artist'))
         self.DefaultModeDropdown.SetValue(artistBackgrounds.get('DefaultMode', 'blend'))
@@ -245,6 +251,7 @@ class MoodsPanel(wx.Panel):
     def updateArtistBackgroundControls(self):
         self.DefaultOpacityLabel.SetLabel(str(self.DefaultOpacitySlider.GetValue()) + '%')
         enabled = self.EnableArtistBackgrounds.GetValue()
+        self.UseCoverArtBackgrounds.Enable(enabled)
         self.MatchFieldDropdown.Enable(enabled)
         self.FallbackFieldDropdown.Enable(enabled)
         self.DefaultModeDropdown.Enable(enabled)
@@ -254,6 +261,7 @@ class MoodsPanel(wx.Panel):
     def OnArtistBackgroundSettingsChanged(self, event):
         artistBackgrounds = self.BeamSettings.getArtistBackgrounds()
         artistBackgrounds['Enabled'] = 'True' if self.EnableArtistBackgrounds.GetValue() else 'False'
+        artistBackgrounds['UseCoverArt'] = 'True' if self.UseCoverArtBackgrounds.GetValue() else 'False'
         artistBackgrounds['MatchField'] = self.MatchFieldDropdown.GetValue()
         artistBackgrounds['FallbackField'] = self.FallbackFieldDropdown.GetValue()
         artistBackgrounds['DefaultMode'] = self.DefaultModeDropdown.GetValue()