+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes in this fork are documented in this file.
 
 ### Added
 
+- Add a feature that allows using Cover Art/ Album Art as the Artist background
+  - coverArt background take the default mode for "blend/replace" and the default opacity for displaying.
+  - if specific background artist is defined, will overrule coverArt brackground
 - **JRiver**: reads singer custom tag from jriver module, reads composer
 - Add an option to define a uniform color for the background instead of a picture.